+++ pararealml/core/operators/pidon/differentiator.py
